+++ src/Compute/Compute.Test/ScenarioTests/AvailabilitySetTests.ps1
@@ -231,3 +231,169 @@ function Test-AvailabilitySetVM
         Clean-ResourceGroup $rgname
     }
 }
+
+<#
+.SYNOPSIS
+Test Availability Set Migration to VMSS Flex
+Note: This test requires the subscription to be enabled for the feature flag Microsoft.Compute/MigrateToVmssFlex
+#>
+function Test-AvailabilitySetMigration
+{
+    param ($loc)
+    # Setup
+    $rgname = Get-ComputeTestResourceName
+
+    try
+    {
+        # Common
+        if ($loc -eq $null)
+        {
+            $loc = Get-ComputeVMLocation;
+        }
+        New-AzResourceGroup -Name $rgname -Location $loc -Force;
+
+        # Create Availability Set
+        $asetName = 'aset' + $rgname;
+        New-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName -Location $loc -Sku 'Aligned' -PlatformFaultDomainCount 2 -PlatformUpdateDomainCount 7;
+        $aset = Get-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName;
+        Assert-NotNull $aset;
+
+        # Create a VM in the Availability Set
+        $vmname = 'vm' + $rgname;
+        $vm = New-TestVmInAvailabilitySet -ResourceGroupName $rgname -Location $loc -AvailabilitySetId $aset.Id -VmName $vmname;
+        $a = $vm | Out-String;
+        Write-Verbose("Get-AzVM output:");
+        Write-Verbose($a);
+        Assert-NotNull $a
+
+        # Create a Flexible VMSS for migration target
+        $vmssName = 'vmss' + $rgname;
+        $vmssConfig = New-AzVmssConfig -Location $loc -OrchestrationMode 'Flexible' -PlatformFaultDomainCount 2 -SinglePlacementGroup $false;
+        $vmss = New-AzVmss -ResourceGroupName $rgname -VMScaleSetName $vmssName -VirtualMachineScaleSet $vmssConfig;
+        Assert-NotNull $vmss;
+        $vmssId = $vmss.Id;
+
+        # Test Validate Migration cmdlet
+        $validateResult = Test-AzAvailabilitySetMigration -ResourceGroupName $rgname -Name $asetName -VirtualMachineScaleSetFlexibleId $vmssId;
+        Assert-NotNull $validateResult;
+
+        # Test StartMigration cmdlet
+        $migrationResult = Start-AzAvailabilitySetMigration -ResourceGroupName $rgname -Name $asetName -VirtualMachineScaleSetFlexibleId $vmssId;
+        Assert-NotNull $migrationResult;
+
+        # Migrate VM to VMSS Flex
+        $migratedVM = Move-AzVirtualMachineToVmss -Id $vm.Id
+        Assert-NotNull $migratedVM;
+
+        # Test Convert cmdlet (creates a new VMSS)
+       # $newVmssName = 'vmss2' + $rgname;
+       # $convertResult = Convert-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName -VirtualMachineScaleSetName $newVmssName;
+       # Assert-NotNull $convertResult;
+
+        Write-Host "Availability Set Migration cmdlets test completed successfully";
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Test Availability Set Convert to VMSS Flex
+Note: This test requires the subscription to be enabled for the feature flag Microsoft.Compute/ConvertToVmssFlex
+#>
+function Test-AvailabilitySetConvert
+{
+    param ($loc)
+    # Setup
+    $rgname = Get-ComputeTestResourceName
+
+    try
+    {
+        # Common
+        if ($loc -eq $null)
+        {
+            $loc = Get-ComputeVMLocation;
+        }
+        New-AzResourceGroup -Name $rgname -Location $loc -Force;
+
+        # Create Availability Set
+        $asetName = 'aset' + $rgname;
+        New-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName -Location $loc -Sku 'Aligned' -PlatformFaultDomainCount 2 -PlatformUpdateDomainCount 5;
+        $aset = Get-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName;
+        Assert-NotNull $aset;
+
+        # Create a VM in the Availability Set
+        $vmname = 'vm' + $rgname;
+        $vm = New-TestVmInAvailabilitySet -ResourceGroupName $rgname -Location $loc -AvailabilitySetId $aset.Id -VmName $vmname;
+        $a = $vm | Out-String;
+        Write-Verbose("Get-AzVM output:");
+        Write-Verbose($a);
+        Assert-NotNull $a
+
+        # Test Convert cmdlet (creates a new VMSS)
+        $newVmssName = 'vmss2' + $rgname;
+        $convertResult = Convert-AzAvailabilitySet -ResourceGroupName $rgname -Name $asetName -VirtualMachineScaleSetName $newVmssName;
+        Assert-NotNull $convertResult;
+
+        Write-Host "Availability Set Migration cmdlets test completed successfully";
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+# Helper: Create a VM in the specified availability set
+function New-TestVmInAvailabilitySet {
+    param(
+        [Parameter(Mandatory = $true)]
+        [string] $ResourceGroupName,
+        [Parameter(Mandatory = $true)]
+        [string] $Location,
+        [Parameter(Mandatory = $true)]
+        [string] $AvailabilitySetId,
+        [Parameter(Mandatory = $true)]
+        [string] $VmName
+    )
+
+    # VM config assigned to Availability Set
+    $vmSize = 'Standard_DS1_v2';
+    $vmConfig = New-AzVMConfig -VMName $VmName -VMSize $vmSize -AvailabilitySetId $AvailabilitySetId;
+
+    # Network
+    $subnet = New-AzVirtualNetworkSubnetConfig -Name ("subnet" + $VmName) -AddressPrefix "10.0.0.0/24";
+    $vnet = New-AzVirtualNetwork -Force -Name ("vnet" + $VmName) -ResourceGroupName $ResourceGroupName -Location $Location -AddressPrefix "10.0.0.0/16" -Subnet $subnet;
+    $vnet = Get-AzVirtualNetwork -Name ("vnet" + $VmName) -ResourceGroupName $ResourceGroupName;
+    $subnetId = $vnet.Subnets[0].Id;
+
+    $pubip = New-AzPublicIpAddress -Force -Name ("pubip" + $VmName) -ResourceGroupName $ResourceGroupName -Location $Location -AllocationMethod Static -DomainNameLabel ("pubip" + $VmName);
+    $pubip = Get-AzPublicIpAddress -Name ("pubip" + $VmName) -ResourceGroupName $ResourceGroupName;
+
+    $nic = New-AzNetworkInterface -Force -Name ("nic" + $VmName) -ResourceGroupName $ResourceGroupName -Location $Location -SubnetId $subnetId -PublicIpAddressId $pubip.Id;
+    $nic = Get-AzNetworkInterface -Name ("nic" + $VmName) -ResourceGroupName $ResourceGroupName;
+
+    $vmConfig = Add-AzVMNetworkInterface -VM $vmConfig -Id $nic.Id;
+    Assert-AreEqual $vmConfig.NetworkProfile.NetworkInterfaces.Count 1;
+   
+    # Minimal OS setup (test-only)
+    $user = "User$($VmName.Substring(0,4))";
+    $password = $PLACEHOLDER;
+    $securePassword = ConvertTo-SecureString $password -AsPlainText -Force;
+    $cred = New-Object System.Management.Automation.PSCredential ($user, $securePassword);
+    $vmConfig = Set-AzVMOperatingSystem -VM $vmConfig -Windows -ComputerName $VmName -Credential $cred;
+
+    $vmConfig = Set-AzVMSourceImage -VM $vmConfig -publisherName "MicrosoftWindowsServer" -offer "WindowsServer" -skus "2022-datacenter-g2" -version "latest";  
+    #$vmConfig = ($imgRef | Set-AzVMSourceImage -VM $vmConfig);
+
+    Assert-AreEqual $p.StorageProfile.ImageReference.Offer $imgRef.Offer;
+    Assert-AreEqual $p.StorageProfile.ImageReference.Publisher $imgRef.PublisherName;
+    Assert-AreEqual $p.StorageProfile.ImageReference.Sku $imgRef.Skus;
+    Assert-AreEqual $p.StorageProfile.ImageReference.Version $imgRef.Version;
+    
+    New-AzVM -ResourceGroupName $ResourceGroupName -Location $Location -VM $vmConfig;
+    return Get-AzVM -ResourceGroupName $ResourceGroupName -Name $VmName;
+}