+++ src/Compute/Compute/Extension/VirtualMachineScaleSetExtensionBaseCmdlet.cs
@@ -33,11 +33,11 @@ namespace Microsoft.Azure.Commands.Compute
             }
         }
 
-        public IVirtualMachineScaleSetVMsOperations VirtualMachineScaleSetVMsClient
+        public IVirtualMachineScaleSetVMSOperations VirtualMachineScaleSetVMsClient
         {
             get
             {
-                return ComputeClient.ComputeManagementClient.VirtualMachineScaleSetVMs;
+                return ComputeClient.ComputeManagementClient.VirtualMachineScaleSetVMS;
             }
         }