+++ src/Sftp/Sftp/CHANGELOG.md
@@ -19,6 +19,12 @@
 -->
 
 ## Upcoming Release
+* Added `BufferSizeInBytes` parameter to `Connect-AzSftp` cmdlet
+    - Allows users to specify buffer size in bytes for SFTP file transfers using the sftp -B flag
+    - Default value is 262144 (256 KB)
+* Added `StorageAccountEndpoint` parameter to `Connect-AzSftp` cmdlet
+    - Allows users to specify a custom storage account endpoint suffix
+    - Useful for connecting to storage accounts with custom endpoints
 
 ## Version 0.1.0
 * Initial release of Az.Sftp module